+++ bugsnag-android-performance/src/main/kotlin/com/bugsnag/android/performance/internal/Tracer.kt
@@ -3,14 +3,14 @@ package com.bugsnag.android.performance.internal
 import android.os.SystemClock
 import com.bugsnag.android.performance.Span
 
-internal class Tracer : SpanProcessor {
+class Tracer : SpanProcessor {
 
     @Suppress("DoubleMutabilityForCollection") // we swap out this ArrayList when we flush batches
     private var batch = ArrayList<SpanImpl>()
 
     private var lastBatchSendTime = SystemClock.elapsedRealtime()
 
-    internal var sampler = Sampler(1.0)
+    internal var sampler: Sampler = ProbabilitySampler(1.0)
 
     internal var worker: Worker? = null