+++ features/fixtures/mazeracer/app/src/main/java/com/bugsnag/mazeracer/scenarios/AppBackgroundedScenario.kt
@@ -1,5 +1,8 @@
 package com.bugsnag.mazeracer.scenarios
 
+import android.content.Intent
+import android.os.Handler
+import android.os.Looper
 import com.bugsnag.android.performance.BugsnagPerformance
 import com.bugsnag.android.performance.PerformanceConfiguration
 import com.bugsnag.android.performance.internal.InternalDebug
@@ -10,10 +13,13 @@ class AppBackgroundedScenario(
     config: PerformanceConfiguration,
     scenarioMetadata: String
 ) : Scenario(config, scenarioMetadata) {
+
+    private val handler = Handler(Looper.getMainLooper())
+
     init {
         InternalDebug.spanBatchSizeSendTriggerPoint = 100
         // this should be longer than the Mazerunner timeout
-        InternalDebug.workerSleepMs = 90_000L
+        InternalDebug.workerSleepMs = 120_000L
     }
 
     override fun startScenario() {
@@ -21,6 +27,16 @@ class AppBackgroundedScenario(
 
         measureSpan("Span 1") {
             Thread.sleep(1)
+        }
+
+        // we send ourselves to the background, since the Appium action isn't very reliable
+        handler.post {
+            context.startActivity(
+                Intent().apply {
+                    setAction(Intent.ACTION_MAIN)
+                    addCategory(Intent.CATEGORY_HOME)
+                },
+            )
         }
     }
 }