+++ features/fixtures/mazeracer/app/src/main/java/com/bugsnag/mazeracer/scenarios/RetryTimeoutScenario.kt
@@ -11,6 +11,7 @@ class RetryTimeoutScenario(
 ) : Scenario(config, scenarioMetadata) {
     init {
         InternalDebug.spanBatchSizeSendTriggerPoint = 1
+        InternalDebug.workerSleepMs = 50L
         InternalDebug.dropSpansOlderThanMs = 100
     }